% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BoussEigen.R
\name{bouss.eigen}
\alias{bouss.eigen}
\title{An eigenvector implementation of the one-dimensional Dupuit-Bousinesq grounwater equation}
\usage{
bouss.eigen(WellDistance = 65340, Storativity = 0.0075,
  Transmisivity = 71940000, ZoneLengths = c(35600, 5700, 8900, 15800),
  DischargeScaleFactor = 500, RechargeData = ZoneVadoseRecharge,
  GWBypassFlow = 1, InitialEigenState = 0)
}
\arguments{
\item{WellDistance}{Distance the well is from the upper edge of the groundwater zone in metres}

\item{Storativity}{#the groundwater storativity, i.e. the fraction of space in the groundwater that is available for water}

\item{Transmisivity}{#the two dimensional flow rate of water through the groundwater m2 per day}

\item{ZoneLengths}{#the lengths (in metres) of each of the zones}

\item{DischargeScaleFactor}{#discharge to groundwater level response gain factor}

\item{GWBypassFlow}{#The amount of discharge that "slips by" the flow recorder, effectively an offset. It could also be considered to be related to the initial depth to the GW, so it can be a negative number. Ideally calibrate for this using the mean error.}

\item{InitialEigenState}{#This is a pseudo offset for the groundwater level, but is not quite......}

\item{RechargeFileName}{#The csv file with all the data in it. This is the daily timeseries of observed discharge and groundwater level, vadose recharge for each zone, river recharge to groundwater and groundwater pumping in mm}
}
\description{
This function takes input data for a catchment about vadose zone recharge, groundwater pumping, soil, the vadose zone and aquifer properties
and simulates observed streamflow and well levels.
Equations are from:
Bidwell, V., Burbery, L., 2011. Groundwater Data Analysis - quantifying aquifer dynamics. Prepared for Envirolink Project 420-NRLC50 (No. 4110/1). Lincoln Ventures Ltd.
which in turn cites the following, though note that the symbol labels are different.
Bidwell, V.J., Stenger, R., Barkle, G.F., 2008. Dynamic analysis of groundwater discharge and partial-area contribution to Pukemanga Stream, New Zealand. Hydrol. Earth Syst. Sci. 12, 975-987. doi:10.5194/hess-12-975-2008
Sloan, W.T., 2000. A physics-based function for modeling transient groundwater discharge at the watershed scale. Water Resour. Res. 36, 225-241. doi:10.1029/1999WR900221
}
\examples{
fishRecharge<-vadose.recharge(ZoneStorageTime=c(4,3,3,3),AquiferZoneDryFractions = list(c(0.8,0.8,0.8,0.8),c(0,0,0,0),c(0,0,0.193,0.8263)),AquiferZoneIrrigFractions=list(c(0,0,0,0),c(0,0,0,0),c(0,0,0.0069,0.1737)),RiverRechargeFractions=c(1,1,0.958,1),RechargeFileName="GoldenBayLandSurfaceRechargeData.csv")
fishDischarge <- bouss.eigen(WellDistance=66000,ZoneLengths=c(35600,5700,8900,15800),Storativity=0.01,Transmisivity=313632,DischargeScaleFactor=950,RechargeData=fishRecharge,GWBypassFlow=1)
}
\keyword{groundwater,}
\keyword{hydrology}

