\name{wtable}
\alias{wtable}
\alias{pem}
\alias{phi.table}
\alias{assoc.twocont}
\alias{assoc.twocat}
\alias{assoc.catcont}
\alias{assoc.yx}
\alias{darma}
\alias{catdesc}
\alias{condesc}
\alias{ggassoc_phiplot}
\alias{ggassoc_boxplot}
\alias{ggassoc_scatter}
\alias{ggassoc_crosstab}

\title{Deprecated functions}

\description{
These functions have been moved to the \code{descriptio} package. You may check its documentation here :
https://nicolas-robette.github.io/descriptio/
}

\usage{
wtable(...)

pem(...)

phi.table(...)

assoc.twocont(...)

assoc.twocat(...)

assoc.catcont(...)

assoc.yx(...)

darma(...)

catdesc(...)

condesc(...)

ggassoc_phiplot(...)

ggassoc_boxplot(...)

ggassoc_scatter(...)

ggassoc_crosstab(...)
}

\arguments{
  \item{...}{arguments are ignored}
}